chore(servers): move first time setup into own component
